rename user_participant_list to participant_admin_list and move the endpoint into the Participant file since we aren't doing any interactions with user in participant admin at this time
