+++ backend/app/resources/ParticipantEndpoint.py
@@ -2,6 +2,7 @@ import datetime
 
 import flask_restful
 from flask import request, g
+from sqlalchemy import func
 
 from app import RestException, db, auth
 from app.model.participant import Participant
@@ -53,3 +54,22 @@ class ParticipantListEndpoint(flask_restful.Resource):
         participants = db.session.query(Participant).all()
         return self.schema.dump(participants)
 
+
+class ParticipantAdminListEndpoint(flask_restful.Resource):
+    def count_participants(self, relationship):
+        query = db.session.query(Participant).filter(Participant.relationship == relationship)
+        count_q = query.statement.with_only_columns([func.count()]).order_by(None)
+        return query.session.execute(count_q).scalar()
+
+    @auth.login_required
+    @requires_roles(Role.admin)
+    def get(self):
+        participant_list = {
+            'num_self_participants': self.count_participants('self_participant'),
+            'num_self_guardians': self.count_participants('self_guardian'),
+            'num_dependents': self.count_participants('dependent'),
+            'num_self_professionals': self.count_participants('self_professional'),
+            'all_participants': ParticipantSchema(many=True).dump(db.session.query(Participant)
+                                                                  .order_by(Participant.relationship).all(), many=True)
+        }
+        return participant_list
