+++ backend/app/resources/UserAndParticipantEndpoint.py
@@ -1,13 +1,12 @@
 import flask_restful
 from flask import request, g
 from marshmallow import ValidationError
-from sqlalchemy import exc, func
+from sqlalchemy import exc
 
 from app import db, RestException, auth
 from app.model.participant import Participant, Relationship
-from app.model.user import User, Role
+from app.model.user import User
 from app.schema.schema import ParticipantSchema
-from app.wrappers import requires_roles
 
 
 class ParticipantBySessionEndpoint(flask_restful.Resource):
@@ -49,23 +48,3 @@ class ParticipantBySessionEndpoint(flask_restful.Resource):
         except exc.IntegrityError as err:
             raise RestException(RestException.INVALID_OBJECT,
                                 details=load_result.errors)
-
-
-class UserParticipantListEndpoint(flask_restful.Resource):
-    def count_participants(self, relationship):
-        query = db.session.query(Participant).filter(Participant.relationship == relationship)
-        count_q = query.statement.with_only_columns([func.count()]).order_by(None)
-        return query.session.execute(count_q).scalar()
-
-    @auth.login_required
-    @requires_roles(Role.admin)
-    def get(self):
-        participant_list = {
-            'num_self_participants': self.count_participants('self_participant'),
-            'num_self_guardians': self.count_participants('self_guardian'),
-            'num_dependents': self.count_participants('dependent'),
-            'num_self_professionals': self.count_participants('self_professional'),
-            'all_participants': ParticipantSchema(many=True).dump(db.session.query(Participant)
-                                                                  .order_by(Participant.relationship).all(), many=True)
-        }
-        return participant_list
