+++ frontend/src/app/participant-admin/participant-admin.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { ApiService } from '../_services/api/api.service';
 import { Participant } from '../_models/participant';
-import { UserParticipantList } from '../_models/user_participant_list';
+import { ParticipantAdminList } from '../_models/participant_admin_list';
 import { MatSort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material';
 
@@ -12,7 +12,7 @@ import { MatTableDataSource } from '@angular/material';
 })
 export class ParticipantAdminComponent implements OnInit {
   @ViewChild(MatSort, {static: true}) sort: MatSort;
-  userParticipantList: UserParticipantList;
+  userParticipantList: ParticipantAdminList;
   participantDataSource: MatTableDataSource<Participant>;
   displayedColumns: string[] = ['id', 'name', 'user_id', 'relationship', 'percent_complete', 'has_consented', 'last_updated'];
   loading: boolean = true;
@@ -22,9 +22,9 @@ export class ParticipantAdminComponent implements OnInit {
   ) { }
 
   ngOnInit() {
-    this.api.getUserParticipantList().subscribe( upl => {
-      this.userParticipantList = upl;
-      this.participantDataSource = new MatTableDataSource(upl.all_participants[0]);
+    this.api.getParticipantAdminList().subscribe(pal => {
+      this.userParticipantList = pal;
+      this.participantDataSource = new MatTableDataSource(pal.all_participants[0]);
       this.participantDataSource.sort = this.sort;
       this.loading = false;
     })
